Implemented adding at position for Linked list; Added tests; Added classes for the stack and testing class.

+++ main/java/datastructure/exceptions/NotAccessiblePositionException.java
@@ -6,7 +6,7 @@ public class NotAccessiblePositionException extends RuntimeException {
         super("There is no such position in the list.");
     }
 
-    public NotAccessiblePositionException(int position, int size) {
+    public NotAccessiblePositionException(long position, long size) {
         super(String.format("There is no such position %d in the list with size %d", position, size));
     }
 }
